refactor(db): extract additional-domain lifecycle into organizations/domains
Move the additional email-domain lifecycle out of db/organizations/mod.rs
into db/organizations/domains.rs: add_additional_domain,
mark_additional_domain_verified, remove_additional_domain,
get_verification_token, list_additional_domains, the background
maintenance (cleanup_stale_additional_domains, record_recheck_result,
list_all_verified_additional_domains), their error/outcome types,
MAX_ADDITIONAL_DOMAINS, the private helpers (verification-token
generation, cross-org conflict scan, session revocation, OrgCasError),
and 25 tests.

mod.rs (2,908 -> 1,164 lines) keeps org CRUD, subdomain claims,
signing-key persistence, and the cross-module auto-release tests; the
auto-release hook is a child-to-parent call into the private
release_ineligible_subdomain. The fresh_store test fixture is hoisted to
module level so both test suites share it. mod.rs re-exports the public
names, so db:: paths and all callers are unchanged.

Closes #636
